Preliminary support for automatic indentation.
This assumes that the only indented lines are legs of a transaction.
It works best if you put blank lines between transactions.
If you do not use blank lines to end a transaction, it will usually work
anyway, since it forces any line starting with a date to be all the way
left.