Follow include statements when completing accounts
Uses Python script to seek through multiple files for accounts,
following 'include' statements. The list of accounts is cached to
'b:beancount_accounts'. You can specify the root file using
'b:beancount_root', and if this is not set, we assume the current file
is the root. Included files that are not found are silently skipped.

Closes #5.